Changed the data type of position in hash_table_has_element to size_t

+++ hashtable.c
@@ -75,7 +75,7 @@ hash_table_t *hash_table_create(int size) {
 }
 
 int hash_table_has_element(hash_table_t *hash_table, char *string) {
-    unsigned int position = hash_table_get_position(hash_table, string);
+    size_t position = hash_table_get_position(hash_table, string);
     hash_table_list_t *list = hash_table->elements[position];
 
 	//Loop through all the entries in the cell string should be in, return 1 if
